src/main/java/com/onlinebanking/backend/service/impl/UserDetailsServiceImpl.java
package com.onlinebanking.backend.service.impl;

import com.onlinebanking.backend.persistent.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.apache.commons.lang3.StringUtils;
import org.springframework.context.annotation.Primary;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.Objects;

/**
 * The implementation of service used to query user details during login.
 *
 * @author Eric Opoku
 * @version 1.0
 * @since 1.0
 */
@Slf4j
@Primary
@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
class UserDetailsServiceImpl implements UserDetailsService {

    private final UserRepository userRepository;

    /**
     * Locates the user based on the username. In the actual implementation, the search
     * may possibly be case sensitive, or case insensitive depending on how the
     * implementation instance is configured. In this case, the <code>UserDetails</code>
     * object that comes back may have a username that is of a different case than what
     * was actually requested..
     *
     * @param username the username identifying the user whose data is required.
     *
     * @return a fully populated user record (never <code>null</code>)
     * @throws UsernameNotFoundException if the user could not be found or the user has no
     *                                   GrantedAuthority
     */
    @Override
    public UserDetails loadUserByUsername(final String username) {
        // Ensure that username is not empty or null.
        if (StringUtils.isNotBlank(username)) {
            var storedUser = userRepository.findByUsername(username);
            if (Objects.isNull(storedUser)) {
                LOG.warn("No record found for storedUser with username {}", username);
                throw new UsernameNotFoundException("User with username " + username + " not found");
            }
            return UserDetailsBuilder.buildUserDetails(storedUser);
        }
        return null;
    }
}
